renamed query.js to default-queries.js
