+++ packages/gatsby-source-shopify/src/default-queries.js
